requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
rich>=13.0.0
py-clob-client>=0.17.0
